# openapi.yaml
openapi: 3.1.0
info:
  title: Maru API
  version: 0.1.0
  description: Versioned API for the Maru convention operating platform.
paths:
...
  /api/v1/me/post-edition-corrections:
    get:
      operationId: privacy_list_my_corrections
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostEditionCorrection'
          description: ''
    post:
      operationId: privacy_propose_my_correction
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostEditionCorrectionCreate'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostEditionCorrection'
          description: ''
  /api/v1/me/privacy-export:
    get:
      operationId: privacy_generate_my_export
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/me/privacy-requests:
    get:
      operationId: privacy_list_my_requests
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectRightsRequest'
          description: ''
    post:
      operationId: privacy_create_my_request
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRightsRequestCreate'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectRightsRequest'
          description: ''
...
  /api/v1/organizations/{organization_id}/editions/{edition_id}/privacy/corrections/{correction_id}/decision:
    post:
      operationId: privacy_decide_post_edition_correction
      parameters:
      - in: path
        name: correction_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostEditionCorrectionDecision'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostEditionCorrection'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/privacy/registration-profile-minimization:
    post:
      operationId: privacy_minimize_registration_profile
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationProfileMinimize'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisposalReceipt'
          description: ''
...
  /api/v1/organizations/{organization_id}/privacy-requests:
    get:
      operationId: privacy_list_staff_requests
      parameters:
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffSubjectRightsRequest'
          description: ''
  /api/v1/organizations/{organization_id}/privacy-requests/{privacy_request_id}/transition:
    post:
      operationId: privacy_transition_staff_request
      parameters:
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: privacy_request_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRightsRequestTransition'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffSubjectRightsRequest'
          description: ''
...
components:
  schemas:
...
    ActionEnum:
      enum:
      - begin_identity_check
      - verify_identity
      - complete
      - deny
      type: string
      description: |-
        * `begin_identity_check` - Begin identity check
        * `verify_identity` - Verify identity and begin work
        * `complete` - Complete
        * `deny` - Deny
...
    DisposalReceipt:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        edition_id:
          type:
          - string
          - 'null'
          format: uuid
          readOnly: true
        policy_id:
          type: string
          format: uuid
          readOnly: true
        target_type:
          type: string
          readOnly: true
        target_id:
          type: string
          format: uuid
          readOnly: true
        disposition:
          allOf:
          - $ref: '#/components/schemas/DispositionEnum'
          readOnly: true
        applied_at:
          type: string
          format: date-time
          readOnly: true
        safe_result_code:
          type: string
          readOnly: true
        downstream_receipts:
          readOnly: true
      required:
      - applied_at
      - disposition
      - downstream_receipts
      - edition_id
      - id
      - policy_id
      - safe_result_code
      - target_id
      - target_type
    DispositionEnum:
      enum:
      - delete
      - minimize
      - retain
      type: string
      description: |-
        * `delete` - Delete
        * `minimize` - Minimize
        * `retain` - Retain with legal basis
...
    PostEditionCorrection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization_id:
          type: string
          format: uuid
          readOnly: true
        edition_id:
          type: string
          format: uuid
          readOnly: true
        target_type:
          type: string
          readOnly: true
        target_id:
          type: string
          format: uuid
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/PostEditionCorrectionStatusEnum'
          readOnly: true
        changed_fields:
          readOnly: true
        reason:
          type: string
          readOnly: true
        requested_at:
          type: string
          format: date-time
          readOnly: true
        decided_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        decision_reason:
          type: string
          readOnly: true
      required:
      - changed_fields
      - decided_at
      - decision_reason
      - edition_id
      - id
      - organization_id
      - reason
      - requested_at
      - status
      - target_id
      - target_type
    PostEditionCorrectionCreate:
      type: object
      properties:
        profile_id:
          type: string
          format: uuid
        changed_fields: {}
        reason:
          type: string
          maxLength: 1000
      required:
      - changed_fields
      - profile_id
      - reason
    PostEditionCorrectionDecision:
      type: object
      properties:
        approve:
          type: boolean
        reason:
          type: string
          maxLength: 1000
      required:
      - approve
      - reason
    PostEditionCorrectionStatusEnum:
      enum:
      - proposed
      - approved
      - rejected
      type: string
      description: |-
        * `proposed` - Proposed
        * `approved` - Approved
        * `rejected` - Rejected
...
    RegistrationProfileMinimize:
      type: object
      properties:
        profile_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
      required:
      - policy_id
      - profile_id
...
    StaffSubjectRightsRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        account_id:
          type: string
          format: uuid
          readOnly: true
        account_display_name:
          type: string
          readOnly: true
        account_email:
          type: string
          format: email
          readOnly: true
        organization_id:
          type:
          - string
          - 'null'
          format: uuid
          readOnly: true
        kind:
          allOf:
          - $ref: '#/components/schemas/SubjectRightsRequestKindEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/SubjectRightsRequestStatusEnum'
          readOnly: true
        requested_at:
          type: string
          format: date-time
          readOnly: true
        request_summary:
          type: string
          readOnly: true
        identity_verified_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        completed_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        safe_outcome_summary:
          type: string
          readOnly: true
      required:
      - account_display_name
      - account_email
      - account_id
      - completed_at
      - id
      - identity_verified_at
      - kind
      - organization_id
      - request_summary
      - requested_at
      - safe_outcome_summary
      - status
...
    SubjectRightsRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization_id:
          type:
          - string
          - 'null'
          format: uuid
          readOnly: true
        kind:
          allOf:
          - $ref: '#/components/schemas/SubjectRightsRequestKindEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/SubjectRightsRequestStatusEnum'
          readOnly: true
        requested_at:
          type: string
          format: date-time
          readOnly: true
        request_summary:
          type: string
          readOnly: true
        identity_verified_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        completed_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        safe_outcome_summary:
          type: string
          readOnly: true
      required:
      - completed_at
      - id
      - identity_verified_at
      - kind
      - organization_id
      - request_summary
      - requested_at
      - safe_outcome_summary
      - status
    SubjectRightsRequestCreate:
      type: object
      properties:
        organization_id:
          type:
          - string
          - 'null'
          format: uuid
        kind:
          $ref: '#/components/schemas/SubjectRightsRequestKindEnum'
        summary:
          type: string
          maxLength: 1000
      required:
      - kind
      - summary
    SubjectRightsRequestKindEnum:
      enum:
      - access
      - correction
      - portability
      - restriction
      - objection
      - deletion
      type: string
      description: |-
        * `access` - Access
        * `correction` - Correction
        * `portability` - Portability
        * `restriction` - Restriction
        * `objection` - Objection
        * `deletion` - Deletion
    SubjectRightsRequestStatusEnum:
      enum:
      - received
      - identity_check
      - in_progress
      - completed
      - denied
      type: string
      description: |-
        * `received` - Received
        * `identity_check` - Identity check
        * `in_progress` - In progress
        * `completed` - Completed
        * `denied` - Denied
    SubjectRightsRequestTransition:
      type: object
      properties:
        action:
          $ref: '#/components/schemas/ActionEnum'
        outcome_summary:
          type: string
          maxLength: 1000
      required:
      - action
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid